[C] Use fragments in FileUpload, show image processing message - Resolves WDP-658
