+++ packages/admin/components/composed/user/UserProfileUpdateForm/UserProfileUpdateForm.tsx
@@ -61,7 +61,7 @@ const UserProfileUpdateForm = ({ data, onSuccess, onCancel }: Props) => {
         <Forms.FileUpload
           label="forms.fields.avatar"
           name="avatar"
-          image={fieldData?.avatar?.small}
+          data={fieldData?.avatar}
           clearName="clearAvatar"
         />
       </Forms.Grid>
@@ -114,14 +114,7 @@ const fragment = graphql`
     email
     username
     avatar {
-      small {
-        png {
-          url
-          alt
-          height
-          width
-        }
-      }
+      ...FileUploadFragment
     }
   }
 `;
