+++ packages/admin/components/forms/FileUpload/BaseFileUpload/BaseFileUploadCurrent.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { FileIconFactory } from "@wdp/lib/factories";
 import { useTranslation } from "react-i18next";
-import * as Styled from "./FileUpload.styles";
+import * as Styled from "./BaseFileUpload.styles";
 import { BaseImage } from "components/atomic";
 import { AssetKind } from "types/graphql-schema";
