+++ packages/admin/components/forms/FileUpload/BaseFileUpload/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./BaseFileUpload";
